Divide sections into components
