+++ frontend/global-styles.css
@@ -1,0 +1,6 @@
+body, html {
+    margin: 0;
+    padding: 0;
+    width: 100%;
+    height: 100%;
+}
