+++ frontend/src/components/index.tsx
@@ -1,3 +1,7 @@
 export { default as ContactForm } from "./ContactForm";
 export { default as Footer } from "./Footer";
 export { default as CardRuta } from "./CardRuta";
+export { default as SliderPage } from "./SliderPage";
+export { default as Sitios } from "./Sitios";
+export { default as Rutas } from "./Rutas";
+export { default as MenuApp } from "./MenuApp";
